Extract text wrapping utility
